refactor: move balances component to components directory
