+++ src/components/Balances/TokenSelector/index.tsx
