+++ src/parts/Topbar/index.tsx
@@ -15,7 +15,7 @@ import {
   tickerToCurrencyIdLiteral
 } from '@interlay/interbtc-api';
 
-import { Balances } from './Balances';
+import { Balances } from 'components/Balances';
 import AccountModal from 'parts/AccountModal';
 import InterlayLink from 'components/UI/InterlayLink';
 import
